ref: seperate waveform component
